Rename stacktrace formatter
