+++ tests/Unit/SDK/Common/Util/test_stacktrace_n_more.phpt
@@ -2,7 +2,7 @@
 Frames should only be collapsed iff matching frames of enclosing exception
 --FILE--
 <?php
-use OpenTelemetry\SDK\Common\Util\TracingUtil;
+use OpenTelemetry\SDK\Common\Exception\StackTraceFormatter;
 
 require_once 'vendor/autoload.php';
 
@@ -11,11 +11,11 @@ function create(?Throwable $e = null): Throwable {
 }
 
 $outer = create(create());
-echo TracingUtil::formatStackTrace($outer), "\n", "\n";
+echo StackTraceFormatter::format($outer), "\n", "\n";
 
 $inner = create();
 $outer = create($inner);
-echo TracingUtil::formatStackTrace($outer), "\n", "\n";
+echo StackTraceFormatter::format($outer), "\n", "\n";
 ?>
 --EXPECTF--
 Exception
